--- Sources/VGSCollectSDK/UIElements/Text Field/Validation/Card/String+SwiftLuhn.swift
@@ -1,18 +1,0 @@
-//
-//  String+SwiftLuhn.swift
-//  VGSCollectSDK
-//
-//  Created by Vitalii Obertynskyi on 9/16/19.
-//  Copyright © 2019 Vitalii Obertynskyi. All rights reserved.
-//
-
-import Foundation
-
-public extension String {
-    
-    /// :nodoc:  Returns string with only numbers
-    var numbersString: Self {
-        let numbersOnlyEquivalent = replacingOccurrences(of: "[^0-9]", with: "", options: .regularExpression, range: nil)
-        return numbersOnlyEquivalent.trimmingCharacters(in: CharacterSet.whitespacesAndNewlines)
-    }
-}
